src/test/java/dev/gothickit/zenkit/world/vob/ContainerTest.java
package dev.gothickit.zenkit.world.vob;

import dev.gothickit.zenkit.GameVersion;
import dev.gothickit.zenkit.LogLevel;
import dev.gothickit.zenkit.Logger;
import dev.gothickit.zenkit.Util;
import dev.gothickit.zenkit.capi.ZenKit;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;

class ContainerTest {
	@BeforeAll
	static void beforeAll() {
		ZenKit.load();
		Logger.setDefault(LogLevel.DEBUG);
	}

	@Test
	void load() {
		var vob = new Container(Util.getResource("G2/VOb/oCMobContainer.zen"), GameVersion.GOTHIC_2);
		assertFalse(vob.isLocked());
		assertEquals("", vob.getKey());
		assertEquals("", vob.getPickString());
		assertEquals("ItMi_Gold:35", vob.getContents());
	}
}